+++ lib/src/parser.dart
@@ -153,7 +153,6 @@ class Lexer {
   Lexer() {
     keywords['+'] = TokenType.PLUS;
     keywords['-'] = TokenType.MINUS;
-    keywords['_'] = TokenType.UNMINUS;
     keywords['*'] = TokenType.TIMES;
     keywords['/'] = TokenType.DIV;
     keywords['%'] = TokenType.MOD;